Split storage challenge gate tests
