+++ crates/oasis7_node/src/tests_storage_challenge_gate.rs
@@ -1,5 +1,9 @@
 use super::*;
 
+#[path = "tests_storage_challenge_gate/provider_route_mocks.rs"]
+mod provider_route_mocks;
+use provider_route_mocks::*;
+
 #[test]
 fn runtime_replication_storage_challenge_gate_blocks_on_local_probe_failure() {
     let dir = temp_dir("challenge-gate-local");
@@ -1178,555 +1182,6 @@ fn runtime_replication_storage_challenge_gate_allows_single_match_without_peer_h
     let _ = fs::remove_dir_all(&dir);
 }
 
-#[test]
-fn runtime_replication_storage_challenge_gate_prefers_dht_blob_providers() {
-    let dir = temp_dir("challenge-gate-provider-selection");
-    let network_impl = Arc::new(ProviderAwareTestNetwork::new(
-        dir.clone(),
-        "storage-provider-1",
-    ));
-    let network: Arc<
-        dyn oasis7_proto::distributed_net::DistributedNetwork<WorldError> + Send + Sync,
-    > = network_impl.clone();
-    let dht = Arc::new(TestReplicaMaintenanceDht::new(
-        "storage-provider-1",
-        "storage-provider-1",
-    ));
-    let pos_config = signed_pos_config_with_signer_seeds(
-        vec![PosValidator {
-            validator_id: "node-a".to_string(),
-            stake: 100,
-        }],
-        &[("node-a", 93)],
-    );
-    let config = NodeConfig::new(
-        "node-a",
-        "world-challenge-provider-selection",
-        NodeRole::Sequencer,
-    )
-    .expect("config")
-    .with_tick_interval(Duration::from_millis(10))
-    .expect("tick")
-    .with_pos_config(pos_config)
-    .expect("pos config")
-    .with_auto_attest_all_validators(true)
-    .with_replication(signed_replication_config(dir.clone(), 93));
-    let mut runtime = with_noop_execution_hook(NodeRuntime::new(config)).with_replication_network(
-        NodeReplicationNetworkHandle::new(Arc::clone(&network))
-            .with_dht(dht)
-            .with_local_provider_id("node-a"),
-    );
 
-    runtime.start().expect("start runtime");
-    let advanced = wait_until(Instant::now() + Duration::from_secs(2), || {
-        runtime.snapshot().consensus.committed_height >= 4
-    });
-    let snapshot = runtime.snapshot();
-    let attempts = network_impl.provider_attempts();
-    assert!(
-        advanced,
-        "runtime did not continue committing when provider-aware fetch-blob was available: committed_height={} network_committed_height={} last_error={:?} attempts={attempts:?}",
-        snapshot.consensus.committed_height,
-        snapshot.consensus.network_committed_height,
-        snapshot.last_error
-    );
-
-    assert!(
-        attempts.iter().any(|providers| {
-            providers
-                .iter()
-                .any(|provider| provider == "storage-provider-1")
-        }),
-        "expected storage challenge gate to request fetch-blob with DHT providers, attempts={attempts:?}"
-    );
-
-    assert!(
-        !snapshot
-            .last_error
-            .as_deref()
-            .map(|reason| reason.contains("storage challenge gate"))
-            .unwrap_or(false),
-        "runtime should not report storage challenge gate failure when provider-aware fetch-blob succeeds: {:?}",
-        snapshot.last_error
-    );
-
-    runtime.stop().expect("stop runtime");
-    let _ = fs::remove_dir_all(&dir);
-}
-
-#[test]
-fn runtime_local_replication_publishes_blob_provider_to_dht() {
-    let dir = temp_dir("publish-local-provider");
-    let network: Arc<
-        dyn oasis7_proto::distributed_net::DistributedNetwork<WorldError> + Send + Sync,
-    > = Arc::new(TestInMemoryNetwork::default());
-    let dht = Arc::new(TestReplicaMaintenanceDht::new("peer-local", "peer-local"));
-    let pos_config = signed_pos_config_with_signer_seeds(
-        vec![PosValidator {
-            validator_id: "node-a".to_string(),
-            stake: 100,
-        }],
-        &[("node-a", 94)],
-    );
-    let config = NodeConfig::new(
-        "node-a",
-        "world-publish-local-provider",
-        NodeRole::Sequencer,
-    )
-    .expect("config")
-    .with_tick_interval(Duration::from_millis(10))
-    .expect("tick")
-    .with_pos_config(pos_config)
-    .expect("pos config")
-    .with_auto_attest_all_validators(true)
-    .with_replication(signed_replication_config(dir.clone(), 94));
-    let mut runtime = with_noop_execution_hook(NodeRuntime::new(config)).with_replication_network(
-        NodeReplicationNetworkHandle::new(Arc::clone(&network))
-            .with_dht(dht.clone())
-            .with_local_provider_id("peer-local"),
-    );
-
-    runtime.start().expect("start runtime");
-    let published = wait_until(Instant::now() + Duration::from_secs(2), || {
-        !dht.published_records().is_empty()
-    });
-    assert!(published, "expected local commit to publish blob provider");
-
-    let published_records = dht.published_records();
-    assert!(
-        published_records
-            .iter()
-            .any(|(_, _, provider_id)| provider_id == "peer-local"),
-        "expected published provider id peer-local, got {published_records:?}"
-    );
-
-    runtime.stop().expect("stop runtime");
-    let _ = fs::remove_dir_all(&dir);
-}
-
-#[test]
-fn runtime_replication_storage_challenge_gate_degrades_after_provider_route_unavailable() {
-    let dir = temp_dir("challenge-gate-provider-fallback");
-    let network_impl = Arc::new(ProviderFallbackTestNetwork::new(dir.clone()));
-    let network: Arc<
-        dyn oasis7_proto::distributed_net::DistributedNetwork<WorldError> + Send + Sync,
-    > = network_impl.clone();
-    let dht = Arc::new(TestReplicaMaintenanceDht::new(
-        "storage-provider-1",
-        "node-a",
-    ));
-    let pos_config = signed_pos_config_with_signer_seeds(
-        vec![PosValidator {
-            validator_id: "node-a".to_string(),
-            stake: 100,
-        }],
-        &[("node-a", 97)],
-    );
-    let config = NodeConfig::new(
-        "node-a",
-        "world-challenge-provider-fallback",
-        NodeRole::Sequencer,
-    )
-    .expect("config")
-    .with_tick_interval(Duration::from_millis(10))
-    .expect("tick")
-    .with_pos_config(pos_config)
-    .expect("pos config")
-    .with_auto_attest_all_validators(true)
-    .with_replication(signed_replication_config(dir.clone(), 97));
-    let mut runtime = with_noop_execution_hook(NodeRuntime::new(config)).with_replication_network(
-        NodeReplicationNetworkHandle::new(Arc::clone(&network))
-            .with_dht(dht)
-            .with_local_provider_id("node-a"),
-    );
-
-    runtime.start().expect("start runtime");
-    let advanced = wait_until(Instant::now() + Duration::from_secs(2), || {
-        runtime.snapshot().consensus.committed_height >= 4
-    });
-    let snapshot = runtime.snapshot();
-    let provider_attempts = network_impl.provider_attempts();
-    let generic_attempts = network_impl.generic_attempts();
-    assert!(
-        advanced,
-        "runtime did not continue committing after provider-route degraded mode: committed_height={} network_committed_height={} last_error={:?} provider_attempts={provider_attempts:?} generic_attempts={generic_attempts}",
-        snapshot.consensus.committed_height,
-        snapshot.consensus.network_committed_height,
-        snapshot.last_error
-    );
-    assert!(
-        provider_attempts.iter().any(|providers| {
-            providers
-                .iter()
-                .any(|provider| provider == "storage-provider-1")
-        }),
-        "expected storage challenge gate to try DHT-selected provider before fallback: {provider_attempts:?}"
-    );
-    assert!(
-        generic_attempts == 0,
-        "storage challenge should not fall back to generic lane request"
-    );
-    assert!(
-        !snapshot
-            .last_error
-            .as_deref()
-            .map(|reason| reason.contains("storage challenge gate"))
-            .unwrap_or(false),
-        "runtime should not hard-fail storage challenge gate when provider route is unavailable: {:?}",
-        snapshot.last_error
-    );
-    assert!(
-        snapshot
-            .consensus
-            .storage_challenge_network_degraded_height
-            .is_some(),
-        "provider route unavailability should be observable as degraded: {:?}",
-        snapshot.consensus.storage_challenge_network_degraded_reason
-    );
-
-    runtime.stop().expect("stop runtime");
-    let _ = fs::remove_dir_all(&dir);
-}
-
-#[test]
-fn runtime_replication_storage_challenge_gate_degrades_after_provider_route_not_found() {
-    let dir = temp_dir("challenge-gate-provider-not-found");
-    let network_impl = Arc::new(ProviderNotFoundFallbackTestNetwork::new(dir.clone()));
-    let network: Arc<
-        dyn oasis7_proto::distributed_net::DistributedNetwork<WorldError> + Send + Sync,
-    > = network_impl.clone();
-    let dht = Arc::new(TestReplicaMaintenanceDht::new(
-        "storage-provider-1",
-        "node-a",
-    ));
-    let pos_config = signed_pos_config_with_signer_seeds(
-        vec![PosValidator {
-            validator_id: "node-a".to_string(),
-            stake: 100,
-        }],
-        &[("node-a", 112)],
-    );
-    let config = NodeConfig::new(
-        "node-a",
-        "world-challenge-provider-not-found",
-        NodeRole::Sequencer,
-    )
-    .expect("config")
-    .with_tick_interval(Duration::from_millis(10))
-    .expect("tick")
-    .with_pos_config(pos_config)
-    .expect("pos config")
-    .with_auto_attest_all_validators(true)
-    .with_replication(signed_replication_config(dir.clone(), 112));
-    let mut runtime = with_noop_execution_hook(NodeRuntime::new(config)).with_replication_network(
-        NodeReplicationNetworkHandle::new(Arc::clone(&network))
-            .with_dht(dht)
-            .with_local_provider_id("node-a"),
-    );
-
-    runtime.start().expect("start runtime");
-    let advanced = wait_until(Instant::now() + Duration::from_secs(2), || {
-        runtime.snapshot().consensus.committed_height >= 4
-    });
-    let snapshot = runtime.snapshot();
-    let provider_attempts = network_impl.provider_attempts();
-    let generic_attempts = network_impl.generic_attempts();
-    assert!(
-        advanced,
-        "runtime did not continue committing after provider-route not-found degraded mode: committed_height={} network_committed_height={} last_error={:?} provider_attempts={provider_attempts:?} generic_attempts={generic_attempts}",
-        snapshot.consensus.committed_height,
-        snapshot.consensus.network_committed_height,
-        snapshot.last_error
-    );
-    assert!(
-        provider_attempts.iter().any(|providers| {
-            providers
-                .iter()
-                .any(|provider| provider == "storage-provider-1")
-        }),
-        "expected storage challenge gate to try DHT-selected provider before degraded mode: {provider_attempts:?}"
-    );
-    assert!(
-        generic_attempts == 0,
-        "storage challenge should not retry generic lane after not-found provider response"
-    );
-    assert!(
-        !snapshot
-            .last_error
-            .as_deref()
-            .map(|reason| reason.contains("storage challenge gate"))
-            .unwrap_or(false),
-        "runtime should not hard-fail storage challenge gate when provider returns not found: {:?}",
-        snapshot.last_error
-    );
-    assert!(
-        snapshot
-            .consensus
-            .storage_challenge_network_degraded_height
-            .is_some(),
-        "provider not-found should be observable as degraded: {:?}",
-        snapshot.consensus.storage_challenge_network_degraded_reason
-    );
-
-    runtime.stop().expect("stop runtime");
-    let _ = fs::remove_dir_all(&dir);
-}
-
-#[derive(Clone)]
-struct ProviderLookupFailureDht;
-
-impl proto_dht::DistributedDht<WorldError> for ProviderLookupFailureDht {
-    fn publish_provider(
-        &self,
-        _world_id: &str,
-        _content_hash: &str,
-        _provider_id: &str,
-    ) -> Result<(), WorldError> {
-        Ok(())
-    }
-
-    fn get_providers(
-        &self,
-        _world_id: &str,
-        _content_hash: &str,
-    ) -> Result<Vec<ProviderRecord>, WorldError> {
-        Err(WorldError::NetworkRequestFailed {
-            code: oasis7_proto::distributed::DistributedErrorCode::ErrTimeout,
-            message: "simulated provider lookup timeout".to_string(),
-            retryable: true,
-        })
-    }
-
-    fn put_world_head(
-        &self,
-        _world_id: &str,
-        _head: &WorldHeadAnnounce,
-    ) -> Result<(), WorldError> {
-        Ok(())
-    }
-
-    fn get_world_head(&self, _world_id: &str) -> Result<Option<WorldHeadAnnounce>, WorldError> {
-        Ok(None)
-    }
-
-    fn put_membership_directory(
-        &self,
-        _world_id: &str,
-        _snapshot: &MembershipDirectorySnapshot,
-    ) -> Result<(), WorldError> {
-        Ok(())
-    }
-
-    fn get_membership_directory(
-        &self,
-        _world_id: &str,
-    ) -> Result<Option<MembershipDirectorySnapshot>, WorldError> {
-        Ok(None)
-    }
-
-    fn put_peer_record(
-        &self,
-        _world_id: &str,
-        _record: &SignedPeerRecord,
-    ) -> Result<(), WorldError> {
-        Ok(())
-    }
-
-    fn get_peer_record(
-        &self,
-        _world_id: &str,
-        _peer_id: &str,
-    ) -> Result<Option<SignedPeerRecord>, WorldError> {
-        Ok(None)
-    }
-}
-
-#[derive(Clone, Default)]
-struct ProviderLookupFailureConnectedPeerTrapNetwork {
-    connected_peer_provider_attempts: Arc<Mutex<usize>>,
-}
-
-impl ProviderLookupFailureConnectedPeerTrapNetwork {
-    fn new() -> Self {
-        Self::default()
-    }
-
-    fn connected_peer_provider_attempts(&self) -> usize {
-        *self
-            .connected_peer_provider_attempts
-            .lock()
-            .expect("lock connected peer provider attempts")
-    }
-}
-
-impl oasis7_proto::distributed_net::DistributedNetwork<WorldError>
-    for ProviderLookupFailureConnectedPeerTrapNetwork
-{
-    fn publish(&self, _topic: &str, _payload: &[u8]) -> Result<(), WorldError> {
-        Ok(())
-    }
-
-    fn subscribe(&self, topic: &str) -> Result<NetworkSubscription, WorldError> {
-        Ok(NetworkSubscription::new(
-            topic.to_string(),
-            Arc::new(Mutex::new(HashMap::new())),
-        ))
-    }
-
-    fn request(&self, protocol: &str, _payload: &[u8]) -> Result<Vec<u8>, WorldError> {
-        Err(WorldError::NetworkProtocolUnavailable {
-            protocol: protocol.to_string(),
-        })
-    }
-
-    fn connected_peer_ids(&self) -> Vec<String> {
-        vec!["observer-light-peer".to_string()]
-    }
-
-    fn request_with_providers(
-        &self,
-        protocol: &str,
-        _payload: &[u8],
-        providers: &[String],
-    ) -> Result<Vec<u8>, WorldError> {
-        if providers
-            .iter()
-            .any(|provider_id| provider_id == "observer-light-peer")
-        {
-            *self
-                .connected_peer_provider_attempts
-                .lock()
-                .expect("lock connected peer provider attempts") += 1;
-            return serde_json::to_vec(&super::replication::FetchBlobResponse {
-                found: true,
-                range_offset_bytes: None,
-                range_complete: None,
-                blob: Some(b"observer-light-wrong-blob".to_vec()),
-            })
-            .map_err(|err| WorldError::DistributedValidationFailed {
-                reason: format!("encode trap blob response failed: {err}"),
-            });
-        }
-        Err(WorldError::NetworkProtocolUnavailable {
-            protocol: protocol.to_string(),
-        })
-    }
-
-    fn register_handler(
-        &self,
-        _protocol: &str,
-        _handler: Box<dyn Fn(&[u8]) -> Result<Vec<u8>, WorldError> + Send + Sync>,
-    ) -> Result<(), WorldError> {
-        Ok(())
-    }
-}
-
-#[derive(Clone, Default)]
-struct ProviderRouteFailureGenericTrapNetwork {
-    generic_attempts: Arc<Mutex<usize>>,
-}
-
-impl ProviderRouteFailureGenericTrapNetwork {
-    fn new() -> Self {
-        Self::default()
-    }
-
-    fn generic_attempts(&self) -> usize {
-        *self.generic_attempts.lock().expect("lock generic attempts")
-    }
-}
-
-impl oasis7_proto::distributed_net::DistributedNetwork<WorldError>
-    for ProviderRouteFailureGenericTrapNetwork
-{
-    fn publish(&self, _topic: &str, _payload: &[u8]) -> Result<(), WorldError> {
-        Ok(())
-    }
-
-    fn subscribe(&self, topic: &str) -> Result<NetworkSubscription, WorldError> {
-        Ok(NetworkSubscription::new(
-            topic.to_string(),
-            Arc::new(Mutex::new(HashMap::new())),
-        ))
-    }
-
-    fn request(&self, _protocol: &str, _payload: &[u8]) -> Result<Vec<u8>, WorldError> {
-        *self.generic_attempts.lock().expect("lock generic attempts") += 1;
-        serde_json::to_vec(&super::replication::FetchBlobResponse {
-            found: true,
-            range_offset_bytes: None,
-            range_complete: None,
-            blob: Some(b"generic-route-wrong-blob".to_vec()),
-        })
-        .map_err(|err| WorldError::DistributedValidationFailed {
-            reason: format!("encode generic trap blob response failed: {err}"),
-        })
-    }
-
-    fn connected_peer_ids(&self) -> Vec<String> {
-        vec!["observer-light-peer".to_string()]
-    }
-
-    fn request_with_providers(
-        &self,
-        protocol: &str,
-        _payload: &[u8],
-        _providers: &[String],
-    ) -> Result<Vec<u8>, WorldError> {
-        Err(WorldError::NetworkProtocolUnavailable {
-            protocol: protocol.to_string(),
-        })
-    }
-
-    fn register_handler(
-        &self,
-        _protocol: &str,
-        _handler: Box<dyn Fn(&[u8]) -> Result<Vec<u8>, WorldError> + Send + Sync>,
-    ) -> Result<(), WorldError> {
-        Ok(())
-    }
-}
-
-struct MalformedProviderResponseNetwork;
-
-impl oasis7_proto::distributed_net::DistributedNetwork<WorldError>
-    for MalformedProviderResponseNetwork
-{
-    fn publish(&self, _topic: &str, _payload: &[u8]) -> Result<(), WorldError> {
-        Ok(())
-    }
-
-    fn subscribe(&self, topic: &str) -> Result<NetworkSubscription, WorldError> {
-        Ok(NetworkSubscription::new(
-            topic.to_string(),
-            Arc::new(Mutex::new(HashMap::new())),
-        ))
-    }
-
-    fn request(&self, protocol: &str, _payload: &[u8]) -> Result<Vec<u8>, WorldError> {
-        Err(WorldError::NetworkProtocolUnavailable {
-            protocol: protocol.to_string(),
-        })
-    }
-
-    fn connected_peer_ids(&self) -> Vec<String> {
-        vec!["storage-provider-1".to_string()]
-    }
-
-    fn request_with_providers(
-        &self,
-        _protocol: &str,
-        _payload: &[u8],
-        _providers: &[String],
-    ) -> Result<Vec<u8>, WorldError> {
-        Ok(b"not-json".to_vec())
-    }
-
-    fn register_handler(
-        &self,
-        _protocol: &str,
-        _handler: Box<dyn Fn(&[u8]) -> Result<Vec<u8>, WorldError> + Send + Sync>,
-    ) -> Result<(), WorldError> {
-        Ok(())
-    }
-}
+#[path = "tests_storage_challenge_gate/provider_routes.rs"]
+mod provider_routes;
